+++ CHANGELOG.md
@@ -17,6 +17,7 @@ For information on changes in released versions of Teku, see the [releases page]
  - Log a warning instead of a verbose error if node is syncing while performing sync committee duties
  - Distributions created from the same git commit and docker image will be identical
  - Optimised storage of latest vote information by batching updates
+ - Validator Registration signature integration with external signer.
 
 ### Bug Fixes
  - Fix not rendering emoticons correctly in graffiti when running in a Docker container